+++ src/main/java/com/tdevilleduc/urthehero/core/service/IStoryService.java
@@ -1,6 +1,7 @@
 package com.tdevilleduc.urthehero.core.service;
 
 import com.tdevilleduc.urthehero.core.model.Story;
+import com.tdevilleduc.urthehero.core.model.dto.StoryDTO;
 
 import java.util.List;
 import java.util.Optional;
@@ -11,6 +12,6 @@ public interface IStoryService {
     boolean notExists(Integer storyId);
     Optional<Story> findById(Integer storyId);
     List<Story> findAll();
-    Story createOrUpdate(Story story);
+    StoryDTO createOrUpdate(StoryDTO storyDto);
     void delete(Integer storyId);
 }