+++ src/main/java/com/tdevilleduc/urthehero/core/service/impl/PageService.java
@@ -1,8 +1,10 @@
 package com.tdevilleduc.urthehero.core.service.impl;
 
+import com.tdevilleduc.urthehero.core.convertor.PageConvertor;
 import com.tdevilleduc.urthehero.core.dao.PageDao;
 import com.tdevilleduc.urthehero.core.exceptions.PageNotFoundException;
 import com.tdevilleduc.urthehero.core.model.Page;
+import com.tdevilleduc.urthehero.core.model.dto.PageDTO;
 import com.tdevilleduc.urthehero.core.service.IPageService;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.helpers.MessageFormatter;
@@ -13,18 +15,23 @@ import org.springframework.util.Assert;
 import java.util.List;
 import java.util.Optional;
 
+import static com.tdevilleduc.urthehero.core.constant.ApplicationConstants.CHECK_PAGEID_PARAMETER_MANDATORY;
+import static com.tdevilleduc.urthehero.core.constant.ApplicationConstants.ERROR_MESSAGE_PAGE_DOESNOT_EXIST;
+
 @Slf4j
 @Service
 public class PageService implements IPageService {
 
     @Autowired
     PageDao pageDao;
+    @Autowired
+    PageConvertor pageConvertor;
 
     public boolean exists(final Integer pageId) {
-        Assert.notNull(pageId, "The pageId parameter is mandatory !");
+        Assert.notNull(pageId, CHECK_PAGEID_PARAMETER_MANDATORY);
         Optional<Page> page = pageDao.findById(pageId);
         if (page.isEmpty()) {
-            log.error("La page avec l'id {} n'existe pas", pageId);
+            log.error(ERROR_MESSAGE_PAGE_DOESNOT_EXIST, pageId);
             return false;
         }
         return true;
@@ -35,7 +42,7 @@ public class PageService implements IPageService {
     }
 
     public Optional<Page> findById(final Integer pageId) {
-        Assert.notNull(pageId, "The pageId parameter is mandatory !");
+        Assert.notNull(pageId, CHECK_PAGEID_PARAMETER_MANDATORY);
         return pageDao.findById(pageId);
     }
 
@@ -43,17 +50,18 @@ public class PageService implements IPageService {
         return pageDao.findAll();
     }
 
-    public Page createOrUpdate(Page page) {
-        return pageDao.save(page);
+    public PageDTO createOrUpdate(PageDTO pageDTO) {
+        Page page = pageConvertor.convertToEntity(pageDTO);
+        return pageConvertor.convertToDto(pageDao.save(page));
     }
 
     public void delete(Integer pageId) {
-        Assert.notNull(pageId, "The pageId parameter is mandatory !");
+        Assert.notNull(pageId, CHECK_PAGEID_PARAMETER_MANDATORY);
         Optional<Page> optional = findById(pageId);
         optional
                 .ifPresentOrElse(page -> pageDao.delete(page),
                         () -> {
-                            throw new PageNotFoundException(MessageFormatter.format("La page avec l'id {} n'existe pas", pageId).getMessage());
+                            throw new PageNotFoundException(MessageFormatter.format(ERROR_MESSAGE_PAGE_DOESNOT_EXIST, pageId).getMessage());
                         }
                 );
     }