+++ src/main/java/com/tdevilleduc/urthehero/core/service/impl/PersonService.java
@@ -1,8 +1,10 @@
 package com.tdevilleduc.urthehero.core.service.impl;
 
+import com.tdevilleduc.urthehero.core.convertor.PersonConvertor;
 import com.tdevilleduc.urthehero.core.dao.PersonDao;
 import com.tdevilleduc.urthehero.core.exceptions.PersonNotFoundException;
 import com.tdevilleduc.urthehero.core.model.Person;
+import com.tdevilleduc.urthehero.core.model.dto.PersonDTO;
 import com.tdevilleduc.urthehero.core.service.IPersonService;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.helpers.MessageFormatter;
@@ -13,17 +15,22 @@ import org.springframework.util.Assert;
 import java.util.List;
 import java.util.Optional;
 
+import static com.tdevilleduc.urthehero.core.constant.ApplicationConstants.CHECK_PERSONID_PARAMETER_MANDATORY;
+import static com.tdevilleduc.urthehero.core.constant.ApplicationConstants.ERROR_MESSAGE_PERSON_DOESNOT_EXIST;
+
 @Slf4j
 @Service
 public class PersonService implements IPersonService {
 
     @Autowired
     PersonDao personDao;
+    @Autowired
+    PersonConvertor personConvertor;
 
     public boolean exists(final Integer personId) {
         Optional<Person> person = personDao.findById(personId);
         if (person.isEmpty()) {
-            log.error("La personne avec l'id {} n'existe pas", personId);
+            log.error(ERROR_MESSAGE_PERSON_DOESNOT_EXIST, personId);
             return false;
         }
         return true;
@@ -34,7 +41,7 @@ public class PersonService implements IPersonService {
     }
 
     public Optional<Person> findById(final Integer personId) {
-        Assert.notNull(personId, "The personId parameter is mandatory !");
+        Assert.notNull(personId, CHECK_PERSONID_PARAMETER_MANDATORY);
         return personDao.findById(personId);
     }
 
@@ -42,17 +49,18 @@ public class PersonService implements IPersonService {
         return personDao.findAll();
     }
 
-    public Person createOrUpdate(Person person) {
-        return personDao.save(person);
+    public PersonDTO createOrUpdate(PersonDTO personDto) {
+        Person person = personConvertor.convertToEntity(personDto);
+        return personConvertor.convertToDto(personDao.save(person));
     }
 
     public void delete(Integer personId) {
-        Assert.notNull(personId, "The personId parameter is mandatory !");
+        Assert.notNull(personId, CHECK_PERSONID_PARAMETER_MANDATORY);
         Optional<Person> optional = findById(personId);
         optional
             .ifPresentOrElse(person -> personDao.delete(person),
                 () -> {
-                    throw new PersonNotFoundException(MessageFormatter.format("La personne avec l'id {} n'existe pas", personId).getMessage());
+                    throw new PersonNotFoundException(MessageFormatter.format(ERROR_MESSAGE_PERSON_DOESNOT_EXIST, personId).getMessage());
                 }
         );
     }