+++ src/main/java/com/tdevilleduc/urthehero/core/service/impl/StoryService.java
@@ -1,11 +1,14 @@
 package com.tdevilleduc.urthehero.core.service.impl;
 
+import com.tdevilleduc.urthehero.core.convertor.StoryConvertor;
 import com.tdevilleduc.urthehero.core.dao.StoryDao;
 import com.tdevilleduc.urthehero.core.exceptions.StoryNotFoundException;
 import com.tdevilleduc.urthehero.core.model.Story;
+import com.tdevilleduc.urthehero.core.model.dto.StoryDTO;
 import com.tdevilleduc.urthehero.core.service.IStoryService;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.helpers.MessageFormatter;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 
@@ -13,12 +16,17 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
+import static com.tdevilleduc.urthehero.core.constant.ApplicationConstants.CHECK_STORYID_PARAMETER_MANDATORY;
+import static com.tdevilleduc.urthehero.core.constant.ApplicationConstants.ERROR_MESSAGE_STORY_DOESNOT_EXIST;
+
 
 @Slf4j
 @Service
 public class StoryService implements IStoryService {
 
     private StoryDao storyDao;
+    @Autowired
+    StoryConvertor storyConvertor;
 
     public StoryService(StoryDao storyDao) {
         this.storyDao = storyDao;
@@ -28,7 +36,7 @@ public class StoryService implements IStoryService {
         Assert.notNull(storyId, "The story parameter is mandatory !");
         Optional<Story> story = storyDao.findById(storyId);
         if (story.isEmpty()) {
-            log.error("L'histoire avec l'id {} n'existe pas", storyId);
+            log.error(ERROR_MESSAGE_STORY_DOESNOT_EXIST, storyId);
             return false;
         }
         return true;
@@ -39,7 +47,7 @@ public class StoryService implements IStoryService {
     }
 
     public Optional<Story> findById(final Integer storyId) {
-        Assert.notNull(storyId, "The storyId parameter is mandatory !");
+        Assert.notNull(storyId, CHECK_STORYID_PARAMETER_MANDATORY);
         return storyDao.findById(storyId);
     }
 
@@ -48,17 +56,18 @@ public class StoryService implements IStoryService {
                 .collect(Collectors.toList());
     }
 
-    public Story createOrUpdate(Story story) {
-        return storyDao.save(story);
+    public StoryDTO createOrUpdate(StoryDTO storyDto) {
+        Story story = storyConvertor.convertToEntity(storyDto);
+        return storyConvertor.convertToDto(storyDao.save(story));
     }
 
     public void delete(Integer storyId) {
-        Assert.notNull(storyId, "The storyId parameter is mandatory !");
+        Assert.notNull(storyId, CHECK_STORYID_PARAMETER_MANDATORY);
         Optional<Story> optional = findById(storyId);
         optional
             .ifPresentOrElse(story -> storyDao.delete(story),
                 () -> {
-                    throw new StoryNotFoundException(MessageFormatter.format("L'histoire avec l'id {} n'existe pas", storyId).getMessage());
+                    throw new StoryNotFoundException(MessageFormatter.format(ERROR_MESSAGE_STORY_DOESNOT_EXIST, storyId).getMessage());
                 }
         );
     }