+++ src/main/java/com/tdevilleduc/urthehero/reader/controller/PersonController.java
@@ -2,6 +2,7 @@ package com.tdevilleduc.urthehero.reader.controller;
 
 import com.tdevilleduc.urthehero.core.exceptions.PersonInternalErrorException;
 import com.tdevilleduc.urthehero.core.model.Person;
+import com.tdevilleduc.urthehero.core.model.dto.PersonDTO;
 import com.tdevilleduc.urthehero.core.service.IPersonService;
 import lombok.extern.slf4j.Slf4j;
 import org.slf4j.helpers.MessageFormatter;
@@ -45,20 +46,20 @@ public class PersonController {
   
     @PutMapping
     public @ResponseBody
-    ResponseEntity<Person> createPerson(@RequestBody @NotNull Person person) {
-        if (person.getId() != null && personService.exists(person.getId())) {
-            throw new PersonInternalErrorException(MessageFormatter.format("Une personne avec l'identifiant {} existe déjà. Elle ne peut être créée", person.getId()).getMessage());
+    ResponseEntity<PersonDTO> createPerson(@RequestBody @NotNull PersonDTO personDto) {
+        if (personDto.getId() != null && personService.exists(personDto.getId())) {
+            throw new PersonInternalErrorException(MessageFormatter.format("Une personne avec l'identifiant {} existe déjà. Elle ne peut être créée", personDto.getId()).getMessage());
         }
-        return ResponseEntity.ok(personService.createOrUpdate(person));
+        return ResponseEntity.ok(personService.createOrUpdate(personDto));
     }
 
     @PostMapping
     public @ResponseBody
-    ResponseEntity<Person> updatePerson(@RequestBody @NotNull Person person) {
-        Assert.notNull(person.getId(), () -> {
+    ResponseEntity<PersonDTO> updatePerson(@RequestBody @NotNull PersonDTO personDto) {
+        Assert.notNull(personDto.getId(), () -> {
             throw new PersonInternalErrorException("L'identifiant de la personne passée en paramètre ne peut pas être null");
         });
-        return ResponseEntity.ok(personService.createOrUpdate(person));
+        return ResponseEntity.ok(personService.createOrUpdate(personDto));
     }
 
     @DeleteMapping(value = "/{personId}")