+++ src/test/java/com/tdevilleduc/urthehero/reader/AbstractTest.java
@@ -1,7 +1,6 @@
 package com.tdevilleduc.urthehero.reader;
 
 import org.junit.ClassRule;
-import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.testcontainers.containers.MySQLContainer;
 
@@ -14,15 +13,10 @@ public abstract class AbstractTest {
             .withPassword("test");
 
     @BeforeAll
-    private static void setupMysql() {
+    protected static void setupMysql() {
         mySqlContainer.start();
         System.setProperty("quarkus.datasource.url", mySqlContainer.getJdbcUrl());
         System.setProperty("quarkus.datasource.username", "test");
         System.setProperty("quarkus.datasource.password", "test");
-    }
-
-    @AfterAll
-    private static void tearDownMysql() {
-        mySqlContainer.stop();
     }
 }