+++ src/test/java/com/tdevilleduc/urthehero/reader/controller/PageControllerTest.java
@@ -7,14 +7,23 @@ import org.junit.jupiter.api.Test;
 import static io.restassured.RestAssured.given;
 
 @QuarkusTest
-class PageControllerTest extends AbstractTest {
+public class PageControllerTest extends AbstractTest {
 
     private static final String uriController = "/api/page";
+
     @Test
-    public void testGetPageById() {
+    public void getPageById_thenSuccess() {
         given()
                 .when().get(uriController + "/1")
                 .then()
-                .statusCode(200);;
+                .statusCode(200);
+    }
+
+    @Test
+    public void getAllPages_thenSuccess() {
+        given()
+                .when().get(uriController + "/all")
+                .then()
+                .statusCode(200);
     }
 }