+++ CHANGELOG.md
@@ -6,8 +6,10 @@
 ### Added
 
 - Integration with Docker [#1][1]
+- Add mapping entity to DTO and reverse [#3][3]
 
 [1]: https://github.com/tdevilleduc/urthehero-reader/issues/1
+[3]: https://github.com/tdevilleduc/urthehero-reader/issues/3
 
 ### Changed